--- caiman_napari/ui_files/main_offline_gui_template.ui.autosave
@@ -1,174 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>MainOfflineGUIWidget</class>
- <widget class="QWidget" name="MainOfflineGUIWidget">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>300</width>
-    <height>905</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Form</string>
-  </property>
-  <layout class="QVBoxLayout" name="verticalLayout">
-   <item>
-    <widget class="QLabel" name="label_2">
-     <property name="font">
-      <font>
-       <pointsize>13</pointsize>
-       <weight>75</weight>
-       <bold>true</bold>
-      </font>
-     </property>
-     <property name="text">
-      <string>CaImAn oflfine</string>
-     </property>
-    </widget>
-   </item>
-   <item>
-    <widget class="QPushButton" name="pushButtonOpenMovie">
-     <property name="text">
-      <string>Open Movie</string>
-     </property>
-    </widget>
-   </item>
-   <item>
-    <widget class="QGroupBox" name="groupBox">
-     <property name="title">
-      <string>Parameter Setting</string>
-     </property>
-     <layout class="QHBoxLayout" name="horizontalLayout_10">
-      <item>
-       <widget class="QPushButton" name="pushButtonParamsMCorr">
-        <property name="text">
-         <string>MCorr</string>
-        </property>
-       </widget>
-      </item>
-      <item>
-       <widget class="QPushButton" name="pushButtonParamsCNMF">
-        <property name="text">
-         <string>CNMF</string>
-        </property>
-       </widget>
-      </item>
-      <item>
-       <widget class="QPushButton" name="pushButtonParamsCNMFE">
-        <property name="text">
-         <string>CNMFE</string>
-        </property>
-       </widget>
-      </item>
-     </layout>
-    </widget>
-   </item>
-   <item>
-    <widget class="QLabel" name="label">
-     <property name="text">
-      <string>Items</string>
-     </property>
-    </widget>
-   </item>
-   <item>
-    <layout class="QHBoxLayout" name="horizontalLayout_9">
-     <item>
-      <widget class="QPushButton" name="pushButtonNewBatch">
-       <property name="text">
-        <string>New Batch</string>
-       </property>
-      </widget>
-     </item>
-     <item>
-      <widget class="QPushButton" name="pushButtonOpenBatch">
-       <property name="text">
-        <string>Open Batch</string>
-       </property>
-      </widget>
-     </item>
-    </layout>
-   </item>
-   <item>
-    <widget class="QListWidget" name="listWidgetItems">
-     <item>
-      <property name="text">
-       <string>cnmf: &lt;name_here&gt;</string>
-      </property>
-     </item>
-     <item>
-      <property name="text">
-       <string>mcorr: A</string>
-      </property>
-     </item>
-     <item>
-      <property name="text">
-       <string>cnmf: A</string>
-      </property>
-     </item>
-     <item>
-      <property name="text">
-       <string>cnmf: B</string>
-      </property>
-     </item>
-    </widget>
-   </item>
-   <item>
-    <layout class="QHBoxLayout" name="horizontalLayout_11">
-     <item>
-      <widget class="QPushButton" name="pushButtonStart">
-       <property name="text">
-        <string>Start</string>
-       </property>
-      </widget>
-     </item>
-     <item>
-      <widget class="QPushButton" name="pushButtonStartItem">
-       <property name="text">
-        <string>Start at item</string>
-       </property>
-      </widget>
-     </item>
-    </layout>
-   </item>
-   <item>
-    <layout class="QHBoxLayout" name="horizontalLayout_12">
-     <item>
-      <widget class="QPushButton" name="pushButtonDelItem">
-       <property name="text">
-        <string>Del Item</string>
-       </property>
-      </widget>
-     </item>
-     <item>
-      <widget class="QPushButton" name="pushButtonAbort">
-       <property name="text">
-        <string>Abort</string>
-       </property>
-      </widget>
-     </item>
-    </layout>
-   </item>
-   <item>
-    <widget class="QTextBrowser" name="textBrowserStdOut"/>
-   </item>
-  </layout>
- </widget>
- <tabstops>
-  <tabstop>pushButtonOpenMovie</tabstop>
-  <tabstop>pushButtonParamsMCorr</tabstop>
-  <tabstop>pushButtonParamsCNMF</tabstop>
-  <tabstop>pushButtonParamsCNMFE</tabstop>
-  <tabstop>pushButtonNewBatch</tabstop>
-  <tabstop>pushButtonOpenBatch</tabstop>
-  <tabstop>listWidgetItems</tabstop>
-  <tabstop>pushButtonStart</tabstop>
-  <tabstop>pushButtonStartItem</tabstop>
-  <tabstop>pushButtonDelItem</tabstop>
-  <tabstop>pushButtonAbort</tabstop>
-  <tabstop>textBrowserStdOut</tabstop>
- </tabstops>
- <resources/>
- <connections/>
-</ui>
